added blank check to programming device
previously there was no check that the page you are going to write
to is blank.  this checks that the page is blank before you try
to write.  a write can only clear a bit but can never set a bit so
writing to a non-blank page is never a good idea.